// packages/datadog-api-client-v1/models/SLOHistoryMetrics.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SLOHistoryMetricsSeries } from './SLOHistoryMetricsSeries';
/**
* A `metric` based SLO history response.  This is not included in responses for `monitor` based SLOs.
*/
export declare class SLOHistoryMetrics {
    'denominator': SLOHistoryMetricsSeries;
    /**
    * The aggregated query interval for the series data. It's implicit based on the query time window.
    */
    'interval': number;
    /**
    * Optional message if there are specific query issues/warnings.
    */
    'message'?: string;
    'numerator': SLOHistoryMetricsSeries;
    /**
    * The combined numerator and denominator query CSV.
    */
    'query': string;
    /**
    * The series result type. This mimics `batch_query` response type.
    */
    'resType': string;
    /**
    * The series response version type. This mimics `batch_query` response type.
    */
    'respVersion': number;
    /**
    * An array of query timestamps in EPOCH milliseconds
    */
    'times': Array<number>;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SLOHistoryMetrics;
    static serialize(data: SLOHistoryMetrics): {
        [key: string]: any;
    };
    constructor();
}